Correctly free DTrace resources.
Cython's cdef attributes are not regular class attributes as in Python.
hasattr(object, name) always returns False for a cdef attribute
corresponding to name.

Instead, we should simply check a pointer's value. I couldn't confirm
it in any documentation but, based on examples I found and my
experiments, it seems that a cdef pointer is always initially set to
NULL, at least when declared as a class attribute.